The Grok cases report under their own name, in their own module
They rode `TestCodexLiveProfiles` as a mixin because a new top-level
class in profile_cases.py would be collected by nothing: run_tests.py
shards `tests/test*.py` alone, so a case module reaches the suite only
through the aggregator import block. Registering the class there is the
whole cost, and it buys back the name -- a broken Grok refusal read as a
Codex failure. The move also drops profile_cases.py from 510 lines to
404, under the 500-line test warning, with room for the case the next
slice needs.
